modulated resolvers
