+++ package.json
@@ -25,6 +25,7 @@
     "@emotion/css": "^11.11.2",
     "@emotion/react": "^11.11.4",
     "@emotion/styled": "^11.11.5",
+    "date-fns": "^3.6.0",
     "emotion": "^11.0.0",
     "graphql": "^16.8.1",
     "kheiron-ui": "^1.0.7",
