+++ src/pages/business/routes/Active/index.tsx
@@ -1,45 +1,7 @@
-import { AtomParagraph, AvatarKUI, TableKUI, UserProfileCardKUI } from 'kheiron-ui';
-import { ActiveTableData, DateType, ProductType, StatusType, UserType } from './type';
+import { TableKUI } from 'kheiron-ui';
+import { ActiveTableData } from 'types';
 import { dataTable } from './mock';
-import { UserCardStyle } from './styles';
-
-const UserResolver = ({ user }: { user: any }) => {
-  const __user = user as UserType;
-  return (
-    <UserProfileCardKUI
-      css={UserCardStyle}
-      image={__user.image}
-      head={__user.name}
-      body={`${__user.email}`}
-    />
-  );
-};
-UserResolver.isUser = function (key: string) {
-  return ['client', 'seller'].includes(key as UserKeysType);
-};
-
-const StatusResolver = ({ status }: { status: any }) => {
-  const __status = status as StatusType;
-  return <AtomParagraph>{__status}</AtomParagraph>;
-};
-
-const ProductsResolver = ({ products }: { products: any }) => {
-  const __products = products as ProductType;
-  console.log(__products);
-  return (
-    <div>
-      {/* <AvatarKUI imgSrc={__products.image} altText={__products.name} /> */}
-      {/* <AtomParagraph>{__products.id}</AtomParagraph> */}
-    </div>
-  );
-};
-
-const DateResolver = ({ date }: { date: any }) => {
-  const __date = date as DateType;
-  return <AtomParagraph>{__date.toDateString()}</AtomParagraph>;
-};
-
-type UserKeysType = 'client' | 'seller';
+import { DateResolver, ProductResolver, StatusResolver, UserResolver } from './resolvers';
 
 type ActiveTableProps = {
   data: ActiveTableData[];
@@ -48,25 +10,25 @@ type ActiveTableProps = {
 const ActiveTable = ({ data }: ActiveTableProps) => {
   return (
     <TableKUI
-      keys={['date', 'products', 'client', 'details', 'status', 'seller']}
+      keys={['date', 'product', 'client', 'details', 'status', 'seller']}
       head={{
         client: 'Client',
         seller: 'Seller',
-        products: 'Products',
+        product: 'Products',
         date: 'Date',
         details: 'Details',
         status: 'Status',
       }}
       body={data as Record<string, any>[]}
       renderBody={({ key, value }) => {
-        if (['client', 'seller'].includes(key as UserKeysType)) {
+        if (UserResolver.isUser(key)) {
           return <UserResolver user={value} />;
         }
         if (key === 'status') {
           return <StatusResolver status={value} />;
         }
-        if (key === 'products') {
-          return <ProductsResolver products={value} />;
+        if (key === 'product') {
+          return <ProductResolver product={value} />;
         }
         if (key === 'date') {
           return <DateResolver date={value} />;
