+++ src/pages/business/routes/Active/resolvers/ProductResolver/index.tsx
@@ -1,0 +1,13 @@
+import { AtomParagraph, AvatarKUI } from 'kheiron-ui';
+import { AvatarStyle } from '../../styles';
+import { ProductType } from 'types';
+
+export const ProductResolver = ({ product }: { product: any }) => {
+  const __product = product as ProductType;
+  return (
+    <div css={AvatarStyle}>
+      <AvatarKUI imgSrc={__product.image} altText={__product.name} />
+      <AtomParagraph>{__product.id}</AtomParagraph>
+    </div>
+  );
+};
