+++ src/pages/business/routes/Active/resolvers/UserResolver/styles.ts
@@ -1,0 +1,11 @@
+import { css } from '@emotion/react';
+
+export const UserCardStyle = css`
+  .KUI-user-profile-card-root {
+    width: auto;
+  }
+  .KUI-atom-element-image {
+    width: 2rem;
+    height: 2rem;
+  }
+`;
