+++ src/pages/business/routes/Active/styles.ts
@@ -1,10 +1,7 @@
 import { css } from '@emotion/react';
 
-export const UserCardStyle = css`
-  .KUI-user-profile-card-root {
-    width: auto;
-  }
-  .KUI-atom-element-image {
+export const AvatarStyle = css`
+  & .KUI-atom-element-image {
     width: 2rem;
     height: 2rem;
   }
